Small typo in documentation
There was a little typo: ou -> or
I also added the world non-coplanar since correspondences are likely and expected to be not on the same plane, otherwise it creates degeneracy which makes it impossible to correctly calculate F matrix.
Minimum number correspondences required is 7. But since it was already 8 ou 7  i changed it to 8 or 7.